grpc/requirements.txt
anyio==3.3.4
asgiref==3.4.1
click==8.0.3
fastapi==0.70.0
Flask==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-tools==1.41.1
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
protobuf==3.19.0
pydantic==1.8.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
Werkzeug==2.0.2

py~=1.11.0
pytest~=7.1.3
pip~=22.2.2
attrs~=22.1.0
wheel~=0.37.1
pluggy~=1.0.0
iniconfig~=1.1.1
requests~=2.28.1
pyparsing~=3.0.9
setuptools~=60.2.0
packaging~=21.3